[udp6] simplify handling of backbone sockets
This commit updates how backbone sockets are tracked. The
`SocketHandle` now tracks the associated `NetifIdentifier`, which is
used to determine if the socket is a backbone socket. The
`SocketHandle::Matches(const MessageInfo &aMessageInfo)` method now
selects backbone or other sockets based on the result of
`aMessageInfo.IsHostInterface()`.

This simplifies the maintenance of backbone sockets in the linked
list, eliminating the need to track multiple heads into the list and
ensuring backbone sockets are always at the end of the list.