[routing-manager] simplify `RxRaTracker::HandleRouterTimer()`
This commit simplifies `RxRaTracker::HandleRouterTimer()`. When all NS
probes to a router fail and it is marked as unreachable,
`HandleRouterTimer()` now directly removes/deprecates its
route/on-link prefixes, replacing a separate method previously used
for this purpose.

Additionally, a new helper method, `ShouldCheckReachability()`, is
added to check if a reachability check (sending NS probes) is needed.
This check is performed only if the router is not already marked as
unreachable and is not the local device itself.