[mesh-forwarder] anycast forwarding to use a child RLOC16 dest
This commit updates `AnycastRouteLookup()` to always select the best
RLOC16 destination for service ALOCs or DHCPv6/ND agent ALOCs, as
discovered from Network Data entries. It removes previous logic that
could choose the parent of a child device as the destination if the
child was selected as the best RLOC16.