[mle] replace `Mle::GetNextHop()` with `GetParentRloc16()`
This commit removes the `Mle::GetNextHop()` method, which returned the
parent RLOC16 (when the parent state was valid) irrespective of a
given destination.

The `Mle::GetNextHop()` method was only used on MTD builds and from
`MeshForwarder::UpdateIp6Route()` method, where
`mle.GetNextHop(Mac::kShortAddrBroadcast)` served as an indirect way
to get the parent RLOC16. This is now replaced with a direct
`GetParentRloc16()` call, clarifying the intended purpose.

On FTD builds, the `MleRouter::GetNextHop()` is used instead of
`Mle::GetNextHop()`. The `MleRouter` method remains unchanged. Since
it would also return the parent's RLOC16 when the device is acting as
a child, using `GetParentRloc16()` covers this case as well.