[mesh-forwarder] update forwarding messages to ALOC destination
This commit updates how messages with an ALOC destination are
forwarded. If the chosen ALOC destination is a child, its parent is
used as the mesh destination unless the device itself is the parent.
Additionally, if the selected ALOC destination is a sleepy child of
the device, the parent node ensures to prepare the message for
indirect transmission. These changes ensure correct forwarding to ED
or SED devices when they register a service and want to receive
messages on their service ALOC.

This commit also adds a new test, `test-030-anycast-forwarding.py`,
which validates the forwarding to ALOC addresses on both ED and SED
devices.