[routing-manager] indicate peer BRs in discovered infra-if routers
This commit updates the public API `otBorderRoutingGetNextRouterEntry`
and `otBorderRoutingRouterEntry` structure to indicate whether a
discovered router on an infrastructure link is likely a peer Thread
Border Router (BR) connected to the same Thread mesh. The related CLI
commands are also updated.

Additionally, this commit adds new tests to validate the discovery and
tracking of peer BRs.