[ot-ctl] add timeout paremater to ot-ctl
When executing the `ot-ctl` with the cli command, the `ot-ctl`
exits after it receves the message `Done` or `Error`. When
executing the `ot-ctl` without any cli commands, the `ot-ctl`
won't exit unless it receives a signal to exit.

OpenThread cli contains the async commands, such as
`linkmetrics query fe80:0:0:0:3092:f334:1455:1ad2 forward 1`, the
result is output after the cli command outputs `Done`. In this case,
the `ot-ctl` with cli command can't catch the command result. If
the users are human, they can execute the 'ot-ctl' without cli
commands to catch the output and then manually abort the `ot-ctl`.
But if the user is script, it is very inconvenient for a script to
send a signal to the `ot-ctl` to exit it.

This commit adds a timeout parameter to the ot-ctl to exit it after
the given time is passed.