[network-data] refactor anycast dest lookup to `NetworkData::Leader` (#10493)
This commit refactors the code and methods responsible for looking up
anycast destination, moving them from the `MeshForwarder` class to
the more appropriate `NetworkData::Leader` class. This better aligns
the responsibilities of each module (e.g. `RouteLookup()` is provided
by `NetworkData::Leader). This is a pure refactor with no changes or
enhancements to the existing implementation.