[cli] add support for CoAP URI queries in CLI requests (#10003)
New feature:

Added CoAP URI query processing functionality to CLI.

Solution:

URI string is checked for a '?' character. If found, the URI query
related part is moved to another string. Afterwards, the modified
`coapUri` string will contain only the path-related parts, and a
second string (`coapUriQuery`) will hold only the URI query related
parts. Then `coapUri` is passed to the original
`otCoapMessageAppendUriPathOptions()` method and `coapUriQuery` is
passed to a new method called `otCoapMessageAppendUriQueryOptions()`,
which - similarly to the other - splits the URI query part into more,
smaller pieces by the delimiter '&' and adds them as separate options
to the CoAP message making the request generated by this CLI function
RFC compliant.  If '?' is not found, everything is processed the old
way.