[cli] update `ping` output to use comma after IPv6 address
This commit updates the CLI `ping` command output. When information
about a received echo response is outputted, a comma (`,`) is now
used after the sender's IPv6 address instead of a colon (`:`). This
change aligns the output with common `ping` outputs (e.g., on Linux)
and avoids ambiguity in parsing the IPv6 address, which itself
contains colons.