[routing-manager] check reachability & send ICMPv6 unreach err
This commit introduces a mechanism  to check for reachability of
messages forwarded by the BR and send an ICMPv6 Destination
Unreachable error to the sender if needed.

If the Border Router (BR) decides to forward an IPv6 message to a
destination outside the Adjacent Infrastructure Link (AIL),  and the
message's source address matches a BR-generated ULA OMR prefix
(with low preference) an ICMPv6 Destination Unreachable message is
sent to the sender.

This check is performed only when a local, non-infrastructure-derived
ULA OMR prefix is published and the BR also publishes a `::/0` route
(due to discovered PIO/RIO prefixes). Under these conditions, a
Thread mesh device might attempt to reach addresses beyond the local
AIL (e.g., the global internet), which would be unreachable.

This feature is controlled by an OT config flag, enabled by default.
Alternatively, this functionality may be implemented within the
platform layer, in which case the configuration should be disabled.

This commit also adds a test case `test-504-br-icmp-unreach-err.py`
validating the newly added behavior.