[srp] introduce `Srp::Coder`
This commit introduces `Srp::Coder`, to encode an SRP message into a
compact, compressed format, reducing the message size. On the server,
the received coded message is decoded to reconstruct an exact replica
of the original message.

The SRP protocol uses DNS resource record format to convey information
about registered services and host addresses. The use of these
records and adherence to the DNS Update general format introduces
overhead and can lead to larger SRP message sizes, which is
undesirable for Thread network. The `Srp::Coder` aims to optimize the
message, resulting in much smaller messages containing all the same
information reducing overall protocol overhead.

This commit updates both SRP client and server implementations in
OpenThread to utilize the SRP coder. New APIs are added to control
the behavior of the SRP client (whether it should prepare and send
coded messages or not). Additionally, public APIs are added to enable
external SRP server implementations to utilize the `Srp::Coder` for
decoding received coded SRP messages.

The SRP client will employ message coding if its use is explicitly
enabled(via API or configuration option), if auto-start mode is
active(allowing the client to parse Thread Network Data and select a
server), and if the Network Data entry corresponding to the selected
server indicates its support for the decoding logic.

The SRP client implements a safeguard fallback mechanism: after a
series of consecutive registration failures with a chosen server, it
will cease sending coded messages. This protects against potential
issues with an SRP server's decoding implementation.

This commit also includes extensive tests covering the behavior of the
newly added `Srp::Coder`:
- A new unit test, `test_srp_coder`, is added to validate specific
  functions/methods of `Srp::Coder`.
- `test_srp_server` and `test_srp_adv_proxy` tests are updated to run
  all test cases twice: with and without the use of the SRP coder.
- Similarly relevant `tests/scripts/thread-cert/test_srp_*` tests are
  updated to run twice, with and without the coder on the client.
- New specific test cases are added in `test_srp_server` to validate
  the format of encoded messages, covering various edge cases.
- The `test_srp_server` also includes a test to validate the SRP
  client's fallback mechanism.