[spinel] stop Spinel time sync when Thread satck is not running
The radio_spinel periodically send Spinel command
SPINEL_PROP_RCP_TIMESTAMP to RCP to synchronize time between host
and RCP. Even if Thread stack is not running, it will periodically
wake up the Thread host.

This commit stops the time sync when the Thread stack is not running
to save host power.