[posix] allow custom implementation of `otPlatInfraIfDiscoverNat64Prefix` (#10566)
This commit moves the macro guard
`OPENTHREAD_POSIX_CONFIG_INFRA_IF_ENABLE` in the code so that it can
allow custom implementation of `otPlatInfraIfDiscoverNat64Prefix` when
`OPENTHREAD_POSIX_CONFIG_INFRA_IF_ENABLE` is disabled.

The use case is on Android platform. We'll not let `ot-daemon`
proactively discover the prefixes by DNS. Instead, System Server will
do the discovery of NAT64 prefix on AIL and it will notify `ot-daemon`
when the AIL NAT64 prefix is discovered/removed.