[srp-client] update `MsgInfo` to include `mMessage` (#10577)
This commit updates and renames the `Info` class to `MsgInfo`, which
now includes `mMessage` as an `OwnedPtr` to the message to be prepared.
This simplifies the code by encapsulating the `Message` and all its
associated data needed to prepare an SRP update message into one data
structure. This structure can be passed as a single input to
different methods preparing various SRP update components, instead of
as two separate inputs.  The OwnedPtr ensures automatic freeing of the
message upon any error.