use number instead of giving backbone_network name when defining TOPOLOGY
when running cert suite, multiple tests are running in parallel, the
extract backbone name is allocated by cert suite script based on
PORT_OFFSET.

For example, if PORT_OFFSET is 10, the default backbone network name is
"backbone10". This commit targets to defineing multiple backbone
networks inside a single tests. For examples, in test_multi_ail.py, if
PORT_OFFSET is 10, then "backbone10.0" and "backbone10.1" are created.

Also, then subnet prefix of the network format is chagnes to:
 backbone10.0 -> 9100:0::0/64 (9100::0/64)
 backbone10.1 -> 9100:1::0/64