[mle] include Supervision TLV only from sleepy child
This commit updates the code to include the Supervision Interval TLV
in MLE messages sent from a child only when the child mode indicates
it is sleepy (`!IsRxOnWhenIdle()`). The Network Diagnostic Child TLV
is also updated to use zero for the "Supervision Interval" field
when the child is not sleepy. The `test-023-mesh-diag.py` is updated
to validate this new behavior.

The parent node still tracks the received supervision interval from a
child in the `Child` entry. This parameter indicates the interval
that would be used if the child were to be supervised.