[core] valgrind reported memory access bugs (#9833)
* posix: check for nlmsg error tlv attributes

if we couldn't set NETLINK_EXT_ACK, there's no extra nlmsg attributes in
the error. avoid UB and walking uninitialized memory by checking the
flag for those attributes. for us, this avoids segfaults and in one
instance, an infinite loop while walking the non-existant attributes.

Signed-off-by: Nick Owens <nick.owens@eero.com>

* posix: zero initialize sigaction struct before use

this removes a valgrind warning about use of uninitialized memory in a
syscall when backtrace is enabled.

Signed-off-by: Nick Owens <nick.owens@eero.com>

* key_manager: zero initialize otSecurityPolicy

valgrind reports that otSecurityPolicy is used uninitialized, so just
make it zero.

clear all bytes when setting to default

Signed-off-by: Nick Owens <nick.owens@eero.com>