[tcat] feat: add 'adapter' argument and use 'BLEDevice' for scanning
Added '-a', '--adapter' arguments to the bbtc.py arguments parser. This allows the selection of the HCI adapter for the scanning procedure.

According to 'Bleak' documentation of the 'BleakClient' class(https://bleak.readthedocs.io/en/latest/api/client.html#bleakclient-class), it's better to use the 'BLEDevice' object in 'BleakClient' during the object instantiation, therefore it has been changed for the scanning procedure.