[csl] simplify calculating CSL phase
This commit simplifies the calculation of CSL phase. The previous
expression is only correct if the CSL phase is smaller than the CSL
period, which may not be true when the CSL period is updated and the new
CSL sample time is not updated yet, this can happen because the CSL IE
is supposed to be filled in an ISR handler.