[tcat] feat: add dataset clear command
Added 'clear' command to the 'dataset' command tree. This allows to remove all entries in the 'ThreadDataset' object used by the script to store the dataset values.

The reason behind this feature is that in the current implementation of the script, the 'ThreadDataset' object entries are always initialized by 'initial_dataset' when running the script.
No command allows to clear/remove the particular entry, which makes this script unable to send an active dataset to the target device without specific dataset values(custom dataset).
To make this possible, the 'clear' command has been added to the 'dataset' command tree, which removes all entries from the 'ThreadDataset' object and, by using existing commands, sets the desired entries in the 'ThreadDataset' object from scratch.
This enables the script to send custom active dataset values to the target device.