[settings] simplify saving of Border Agent ID
This commit simplifies how Border Agent ID is saved in non-volatile
`Settings` by utilizing generic methods designed for single-value
setting entries. As a result, `Settings::BorderAgentId` now only
needs to define the key and the associated entry value type. This
eliminates the need for `otBorderAgentId` to be defined as packed,
thereby simplifying this structure.